+++ rat-text/src/text_input.rs
@@ -245,13 +245,18 @@ fn render_ref(widget: &TextInput<'_>, area: Rect, buf: &mut Buffer, state: &mut 
         let c = state.cursor();
         let o = state.offset();
 
-        let no = if c < o {
+        let mut no = if c < o {
             c
         } else if c >= o + (state.inner.width + state.dark_offset.0) as upos_type {
             c.saturating_sub((state.inner.width + state.dark_offset.0) as upos_type)
         } else {
             o
         };
+        // correct by one at right margin. block cursors appear as part of the
+        // right border otherwise.
+        if c == no + (state.inner.width + state.dark_offset.0) as upos_type {
+            no = no.saturating_add(1);
+        }
         state.set_offset(no);
     }
 
@@ -715,9 +720,11 @@ impl TextInputState {
         self.value.anchor().x
     }
 
-    /// Set the cursor position, reset selection.
+    /// Set the cursor position.
+    /// Scrolls the cursor to a visible position.
     #[inline]
     pub fn set_cursor(&mut self, cursor: upos_type, extend_selection: bool) -> bool {
+        self.scroll_cursor_to_visible();
         self.value
             .set_cursor(TextPosition::new(cursor, 0), extend_selection)
     }
@@ -736,15 +743,19 @@ impl TextInputState {
     }
 
     /// Selection.
+    /// Scrolls the cursor to a visible position.  
     #[inline]
     pub fn set_selection(&mut self, anchor: upos_type, cursor: upos_type) -> bool {
+        self.scroll_cursor_to_visible();
         self.value
             .set_selection(TextPosition::new(anchor, 0), TextPosition::new(cursor, 0))
     }
 
     /// Selection.
+    /// Scrolls the cursor to a visible position.
     #[inline]
     pub fn select_all(&mut self) -> bool {
+        self.scroll_cursor_to_visible();
         self.value.select_all()
     }
 
@@ -1199,53 +1210,40 @@ impl TextInputState {
     #[inline]
     pub fn move_right(&mut self, extend_selection: bool) -> bool {
         let c = min(self.cursor() + 1, self.len());
-        let c = self.set_cursor(c, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(c, extend_selection)
     }
 
     /// Move to the previous char.
     #[inline]
     pub fn move_left(&mut self, extend_selection: bool) -> bool {
         let c = self.cursor().saturating_sub(1);
-        let c = self.set_cursor(c, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(c, extend_selection)
     }
 
     /// Start of line
     #[inline]
     pub fn move_to_line_start(&mut self, extend_selection: bool) -> bool {
-        let c = self.set_cursor(0, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(0, extend_selection)
     }
 
     /// End of line
     #[inline]
     pub fn move_to_line_end(&mut self, extend_selection: bool) -> bool {
-        let c = self.len();
-        let c = self.set_cursor(c, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(self.len(), extend_selection)
     }
 
     #[inline]
     pub fn move_to_next_word(&mut self, extend_selection: bool) -> bool {
         let cursor = self.cursor();
         let end = self.next_word_end(cursor);
-        let c = self.set_cursor(end, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(end, extend_selection)
     }
 
     #[inline]
     pub fn move_to_prev_word(&mut self, extend_selection: bool) -> bool {
         let cursor = self.cursor();
         let start = self.prev_word_start(cursor);
-        let c = self.set_cursor(start, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(start, extend_selection)
     }
 }
 
@@ -1346,9 +1344,7 @@ impl TextInputState {
 
         let cx = self.screen_to_col(scx);
 
-        let c = self.set_cursor(cx, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(cx, extend_selection)
     }
 
     /// Set the cursor position from screen coordinates,
@@ -1378,9 +1374,7 @@ impl TextInputState {
             }
         }
 
-        let c = self.set_cursor(cursor, extend_selection);
-        self.scroll_cursor_to_visible();
-        c
+        self.set_cursor(cursor, extend_selection)
     }
 
     /// Scrolling